Fixed race conditions for background adjustments
We had a problem where the UI could indefinitely lock up if making adjustments on a small image since updates would happen fast enough that they would be complete before the presenter locked the UI, thus causing the UI to lock forever. Fixed this by changing the condition for which the update to the model is made.