Fixed the slider CPU use crash.
The issue was a thread was being spawned after GTK sent
the signal to indicate the slider was no longer in use.
This caused a thread to sit in an infinite loop and
consume resources.

The fix was to track the use of threads by giving them
optional names.  This list of active threads is only kept
up to date in the slider method, but any method could also
keep it up to date by invoking _prune_active_threads().