Toned down distortions
We had to reduce the extremity of some of the distortions and blurs in the photo app. This mainly amounted to just fiddling with some paramters to the image processing functions. We removed the bulge option, though the code to apply it is still there in case we want to resurrect it someday.