Hash filename + date to get background image filename
Rather than saving the background images in the pictures directory
we now save it in a special directory for background images in the .cache
folder. We hash the filename+current timestamp in order to get
unique filenames for each background image. At some point, we want
to add code that will go in and periodically clean up this directory.