Tempory fix for dropped screen refreshes with MALI
Hopefully very temporary. The photo app will not update the photo
content every other time a filter is applied. This happens only on
MALI but makes the app entirely unusable. Adding a queue_redraw in
an idle_add fixes the problem. The right way to do this is to get
a fix in for MALI or clutter-gtk. When that happens we should
revert.