Use window set_sensitive to lock the UI when updating a photo
Way back when, calling window.set_sensitive(False) was broken and
causing a screen flash. This does not seem to be the case with new
Gtk. Using set sensitive means we can avoid a bunch on manual checks in
the presenter. Also, UI is no longer reactive when photo is updating,
making it more obvious to the user that his clicks won't do anything