Updated transform buttons UI
The rotate and crop buttons in the transform menu were updated
to match new design specs and to be more visible against the
darker background recently added. The background gradient & shadow
are lighter, and the text/icon color when pressed is now the same
shade of grey as during the normal state.

[endlessm/eos-sdk#1364]