Implemented arrows UI.
The old arrows of the expanding sidebar menus have been replaced with the new assets.
Because the assets weren't quite the right size, the images would stretch to match
the frame. Therefore, some CSS tinkering with background sizes was done to make the
arrows match the design specifications as closely as possible.

[endlessm/eos-skd#1350]