photos_left_toolbar: change style class name
We were using the style class "label" which is applied to all gtk
labels, to try to uniquely refer to a label in our category expanders.

I believe as part of the gtk update, styling meant only for the
expanders started applying to buttons inside the expanders.

Switch to a unique style class name to avoid the collision
[endlessm/eos-sdk#3466]