Restore dialog title
GTK doesn't display the title of transient dialogs anymore if they don't
use the header bar. We don't want to use the header bar at this point,
because it would disrupt our design too much, so this packs a fake title
bar with gtk_window_set_titlebar(). It adds a bit of CSS to make the
border between the title bar and the rest of the window a little less
jarring.
[endlessm/eos-sdk#3470]