photos_presenter: defer showing crop options
The photos app some times crashes while opening an image and
then clicking on the transform->crop option. This is hard to
reproduce, approximately 1 in 20 attemps of:

  1. run eos-photos ~/path/to/image_file
  2. click on transforms.
  3. click on crop.

I could not find the exact reason why this happens, but noticed
during my tests that the issue seems to decrease in frequency
when things were slown down while updating the crop options css
styles, to the point where it can be avoided if deferred from
the crop overlay creation.

This workaround defers showing the crop options and avoids the
crash.

With this patch the app does not crash on several consecutive
tests (more than 40). Without this patch, the app crashes on less
than 20 tests.

https://phabricator.endlessm.com/T11249