splash: Refactor useless Gtk.HBoxes
The extra Gtk.HBox around the splash screen label and button fulfilled no
function that couldn't already be fulfilled by halign and valign. (This
was probably not true in GTK 3.6 when this class was originally written.)

I originally hoped this would solve the problem of requesting too much
height on composite, but it didn't. However, it does make the code leaner
and causes no other changes as far as I can see, so I'm submitting it
anyhow.

https://phabricator.endlessm.com/T11401