facebook: Update photo post flow to 2.12 API
Hopefully, this is all that's needed.

One change that may affect us is that apps will need to provide a hash of
the app secret in all their API calls if they have restrictions on the
user's age or the country in which the app can be used. However, we
don't have any restrictions currently, and including the app secret in
the app would be difficult, especially if we want to be building it on
Flathub.

https://phabricator.endlessm.com/T21746